+++ js/custom-setups/card-setup.js
@@ -1,30 +1,13 @@
+import { minMaxes } from "./min-max.js";
+
 export function customCardSetup(s) {
-    if (_.startsWith(s.id, "soccer")) {
-        customSoccerCardSetup(s.id, s.appearance.width, s.appearance.height);
+    customWidthHeightCardSetup(s.id, s.appearance.width, s.appearance.height);
+    if (s.id === "ice-hockey-iihf") {
+        customCornerRadiusSetup(s.id, s.appearance.cornerRadius);
     }
 }
 
-function customSoccerCardSetup(id, width, height) {
-    const minMaxes = {
-        "soccer-ifab-yd": {
-            minWidth: 100,
-            maxWidth: 130,
-            minHeight: 50,
-            maxHeight: 100,
-        },
-        "soccer-ifab-m": {
-            minWidth: 90,
-            maxWidth: 120,
-            minHeight: 45,
-            maxHeight: 90,
-        },
-        "soccer-ncaa": {
-            minWidth: 115,
-            maxWidth: 120,
-            minHeight: 70,
-            maxHeight: 75,
-        },
-    };
+function customWidthHeightCardSetup(id, width, height) {
     const card = d3.select(`#${id}`).attr("href", undefined);
     const dim = card.select(".card-text").select(".dimensions");
     const minMax = minMaxes[id];
@@ -39,10 +22,13 @@ function customSoccerCardSetup(id, width, height) {
         .attr("name", `${id}-width`)
         .attr("min", minMax.minWidth)
         .attr("max", minMax.maxWidth)
-        .attr("value", width);
-    widthField
-        .append("span")
-        .text(`(min: ${minMax.minWidth}, max: ${minMax.maxWidth})`);
+        .attr("value", width)
+        .attr("disabled", minMax.minWidth === minMax.maxWidth ? true : null);
+    if (minMax.minWidth !== minMax.maxWidth) {
+        widthField
+            .append("span")
+            .text(`(min: ${minMax.minWidth}, max: ${minMax.maxWidth})`);
+    }
 
     const heightField = dim.append("div");
     heightField.append("label").attr("for", `${id}-height`).text("Height:");
@@ -53,10 +39,13 @@ function customSoccerCardSetup(id, width, height) {
         .attr("name", `${id}-height`)
         .attr("min", minMax.minHeight)
         .attr("max", minMax.maxHeight)
-        .attr("value", height);
-    heightField
-        .append("span")
-        .text(`(min: ${minMax.minHeight}, max: ${minMax.maxHeight})`);
+        .attr("value", height)
+        .attr("disabled", minMax.minHeight === minMax.maxHeight ? true : null);
+    if (minMax.minHeight !== minMax.maxHeight) {
+        heightField
+            .append("span")
+            .text(`(min: ${minMax.minHeight}, max: ${minMax.maxHeight})`);
+    }
 
     card.select("button").on("click", function () {
         const width = d3.select(`#${id}-width`).property("value");
@@ -64,6 +53,45 @@ function customSoccerCardSetup(id, width, height) {
         let params = new URLSearchParams({
             width: width,
             height: height,
+        });
+        window.location.href = `./${id}?${params.toString()}`;
+    });
+}
+
+function customCornerRadiusSetup(id, cornerRadius) {
+    const minMax = minMaxes[id];
+    const card = d3.select(`#${id}`).attr("href", undefined);
+    const dim = card.select(".card-text").select(".dimensions");
+    const cornerRadiusField = dim.append("div");
+    cornerRadiusField
+        .append("label")
+        .attr("for", `${id}-corner-radius`)
+        .text("Corner Radius:");
+    cornerRadiusField
+        .append("input")
+        .attr("id", `${id}-corner-radius`)
+        .attr("type", "number")
+        .attr("name", `${id}-corner-radius`)
+        .attr("min", minMax.minCornerRadius)
+        .attr("max", minMax.maxCornerRadius)
+        .attr("step", 0.1)
+        .attr("value", cornerRadius);
+    cornerRadiusField
+        .append("span")
+        .text(
+            `(min: ${minMax.minCornerRadius}, max: ${minMax.maxCornerRadius})`
+        );
+
+    card.select("button").on("click", function () {
+        const width = d3.select(`#${id}-width`).property("value");
+        const height = d3.select(`#${id}-height`).property("value");
+        const cornerRadius = d3
+            .select(`#${id}-corner-radius`)
+            .property("value");
+        let params = new URLSearchParams({
+            width: width,
+            height: height,
+            cornerRadius: cornerRadius,
         });
         window.location.href = `./${id}?${params.toString()}`;
     });