+++ js/custom-setups/min-max.js
@@ -1,0 +1,34 @@
+export const minMaxes = {
+    "soccer-ifab-yd": {
+        minWidth: 100,
+        maxWidth: 130,
+        minHeight: 50,
+        maxHeight: 100,
+    },
+    "soccer-ifab-m": {
+        minWidth: 90,
+        maxWidth: 120,
+        minHeight: 45,
+        maxHeight: 90,
+    },
+    "soccer-ncaa": {
+        minWidth: 115,
+        maxWidth: 120,
+        minHeight: 70,
+        maxHeight: 75,
+    },
+    "indoor-lacrosse": {
+        minWidth: 180,
+        maxWidth: 200,
+        minHeight: 85,
+        maxHeight: 85,
+    },
+    "ice-hockey-iihf": {
+        minWidth: 60,
+        maxWidth: 60,
+        minHeight: 26,
+        maxHeight: 30,
+        minCornerRadius: 7,
+        maxCornerRadius: 8.5,
+    },
+};