+++ CHANGELOG.md
@@ -7,6 +7,26 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to the spirit of [Semantic Versioning](https://semver.org/spec/v2.0.0.html) as best it can without a proper API.
 
+## [2.3.0] - 2024-12-26
+
+### Added
+
+#### New Playing Areas
+
+-   Football Net (IFAB)
+-   Soccer Net (NCAA)
+-   Indoor Lacrosse (NLL customizable in feet)
+
+### Fixed
+
+-   Bandage fix for ghost dots not appearing promptly by disabling the size appearance transition for ghost dots.
+-   Fix sizing for default Soccer (NCAA) playing area.
+-   Fix dot transformation when changing dimensions on customizable dimension playing areas.
+
+### Changed
+
+-   Adjusted Ice Hockey (IIHF) playing area to allow customization of width and corner radius.
+
 ## [2.2.0] - 2024-06-26
 
 ### Added
@@ -22,7 +42,7 @@ and this project adheres to the spirit of [Semantic Versioning](https://semver.o
 
 ### Changed
 
--   Adjusted Football (NCAA) playing area to also allow customization of field height and width.
+-   Adjusted Soccer (NCAA) playing area to also allow customization of field height and width.
 
 ## [2.1.0] - 2023-09-30
 
@@ -259,7 +279,8 @@ and this project adheres to the spirit of [Semantic Versioning](https://semver.o
 -   Ability to click on rink to create dot and add row to table with the current details.
 -   Ability to download and upload table.
 
-[2.1.0]: https://github.com/nguyenank/shot-plotter/releases/tag/v2.2.0
+[2.3.0]: https://github.com/nguyenank/shot-plotter/releases/tag/v2.3.0
+[2.2.0]: https://github.com/nguyenank/shot-plotter/releases/tag/v2.2.0
 [2.1.0]: https://github.com/nguyenank/shot-plotter/releases/tag/v2.1.0
 [2.0.1]: https://github.com/nguyenank/shot-plotter/releases/tag/v2.0.1
 [2.0.0]: https://github.com/nguyenank/shot-plotter/releases/tag/v2.0.0